# Port more performant UUID parser