Closes https://github.com/zio/zio-json/issues/193

### Other upgrades
- Upgrade Scala to 2.12.14/2.13.6
- Upgrade ZIO to 1.0.9

### Notes
- I had to port `parseLong` as well since `java.lang.Long.parseLong` is not present under ScalaJS
- UUID parser is a port of https://github.com/openjdk/jdk/commit/ebadfaeb2e1cc7b5ce5f101cd8a539bc5478cf5b#diff-8fd113f7046fdd4617079fd4b61e4957e80119e07189d619bfdfe468b71e99ea
- Credits to @plokhotnyuk for his work