@calvinlfer 2021-06-03
> Do you have benchmark numbers?

I'll run some JMH benchmarks and will post