@calvinlfer 2021-06-03
```scala
package zio.json.uuid

import org.openjdk.jmh.annotations._
import zio.Chunk

import java.util.UUID
import java.util.concurrent.TimeUnit

@State(Scope.Thread)
@Warmup(iterations = 5, time = 1, timeUnit = TimeUnit.SECONDS)
@Measurement(iterations = 5, time = 1, timeUnit = TimeUnit.SECONDS)
@OutputTimeUnit(TimeUnit.MILLISECONDS)
@Fork(value = 1)
class UUIDBenchmarks {
  private[this] var unparsedUUIDChunk: Chunk[String] = null

  @Setup
  def setup(): Unit =
    unparsedUUIDChunk = Chunk.fromIterable(Range(1, 1000).map(_ => UUID.randomUUID().toString))

  @Benchmark
  def parseInBuiltUUID: Chunk[UUID] =
    unparsedUUIDChunk.map(UUID.fromString)

  @Benchmark
  def parseCustom: Chunk[UUID] =
    unparsedUUIDChunk.map(UUIDParser.unsafeParse)
}
```

```
[info] Benchmark                                                      Mode  Cnt      Score       Error   Units
// new
[info] UUIDBenchmarks.parseCustom                                    thrpt    5     39.310 ±     1.617  ops/ms
[info] UUIDBenchmarks.parseCustom:·gc.alloc.rate                     thrpt    5   1000.600 ±    41.218  MB/sec
[info] UUIDBenchmarks.parseCustom:·gc.alloc.rate.norm                thrpt    5  40064.174 ±     1.402    B/op
[info] UUIDBenchmarks.parseCustom:·gc.churn.G1_Eden_Space            thrpt    5   1060.991 ±   860.691  MB/sec
[info] UUIDBenchmarks.parseCustom:·gc.churn.G1_Eden_Space.norm       thrpt    5  42466.206 ± 34098.970    B/op
[info] UUIDBenchmarks.parseCustom:·gc.churn.G1_Old_Gen               thrpt    5      0.016 ±     0.119  MB/sec
[info] UUIDBenchmarks.parseCustom:·gc.churn.G1_Old_Gen.norm          thrpt    5      0.645 ±     4.843    B/op
[info] UUIDBenchmarks.parseCustom:·gc.count                          thrpt    5     13.000              counts
[info] UUIDBenchmarks.parseCustom:·gc.time                           thrpt    5      9.000                  ms

// old
[info] UUIDBenchmarks.parseInBuiltUUID                               thrpt    5     12.563 ±     0.284  ops/ms
[info] UUIDBenchmarks.parseInBuiltUUID:·gc.alloc.rate                thrpt    5    319.824 ±     7.284  MB/sec
[info] UUIDBenchmarks.parseInBuiltUUID:·gc.alloc.rate.norm           thrpt    5  40064.038 ±     0.020    B/op
[info] UUIDBenchmarks.parseInBuiltUUID:·gc.churn.G1_Eden_Space       thrpt    5    316.169 ±   683.542  MB/sec
[info] UUIDBenchmarks.parseInBuiltUUID:·gc.churn.G1_Eden_Space.norm  thrpt    5  39529.271 ± 85497.688    B/op
[info] UUIDBenchmarks.parseInBuiltUUID:·gc.count                     thrpt    5      4.000              counts
[info] UUIDBenchmarks.parseInBuiltUUID:·gc.time                      thrpt    5      9.000                  ms
```

### Specifications
* GraalVM-EE Java 11 21.1.0
* MacBook Pro (16-inch, 2019)
* 2.6 GHz 6-Core Intel Core i7
* 16 GB 2667 MHz DDR4