@calvinlfer 2021-06-04
After applying most of the optimizations that @plokhotnyuk suggested, I see the performance increase from `39 ops/s` to `41-43 ops/s`