@plokhotnyuk 2021-06-05
> After applying most of the optimizations that @plokhotnyuk suggested, I see the performance increase from `39 ops/s` to `41-43 ops/s`

I would add an additional benchmark that have non-standard input length from 9 to 35 characters.